% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateManupulations.R
\name{water_year}
\alias{water_year}
\title{Derive water year value for a date}
\usage{
water_year(date, start_month = 10)
}
\arguments{
\item{date}{the (vector of) dates for which the water year will be calculated}

\item{start_month}{the month in which the water year starts, default is October}
}
\value{
The water year value
}
\description{
Derive water year value for a date
}
\examples{
water_year(as.Date(c("2010-11-03", "2013-02-03")))
}
